Added RepositoryPattern

+++ Domain/Entities/Statistic.cs
@@ -6,5 +6,5 @@ public class Statistic:BaseEntity
 {
     public int Count { get; set; }  
     public string Value { get; set; }  
-    public int DisplayOrder { get; set; }  
+    public short DisplayOrder { get; set; }  
 }
